Open and close the modal programatically
This behavior allows us to pass a custom config to each modal instead of relying on the init() config, and to move away from the data attributes.

This commit also disables closing modals via click outside, because a bug is causing the modal to get closed when clicking inside the modal.